+++ Gruntfile.js
@@ -150,11 +150,10 @@ module.exports = function(grunt) {
   function rewrite(dist) {
     var manifest = grunt.file.readJSON('extension-dist/manifest.json');
     if (dist === 'beta') {
-      manifest.name += ' Beta';
+      manifest.name = 'Destiny Item Manager Beta Shortcut';
     }
     manifest.version = dist === 'beta' ? betaVersion : pkg.version;
 
-    manifest.content_scripts[0].matches = [`https://${dist}.destinyitemmanager.com/*`];
     grunt.file.write('extension-dist/manifest.json', JSON.stringify(manifest));
     var mainjs = grunt.file.read('extension-dist/main.js');
     mainjs = mainjs.replace('localhost:8080', `${dist}.destinyitemmanager.com`);