+++ src/app/destiny2/d2-inventory.component.js
@@ -9,10 +9,16 @@ export const D2InventoryComponent = {
   controller: D2InventoryController
 };
 
-function D2InventoryController($scope, D2StoresService, D2BucketsService) {
+function D2InventoryController($rootScope, $scope, D2StoresService, D2BucketsService) {
   'ngInject';
 
   const vm = this;
+
+  vm.featureFlags = {
+    dnd: $featureFlags.dnd
+  };
+
+  let dragBox;
 
   this.$onInit = function() {
     D2BucketsService.getBuckets().then((buckets) => {
@@ -25,5 +31,18 @@ function D2InventoryController($scope, D2StoresService, D2BucketsService) {
 
   $scope.$on('dim-refresh', () => {
     D2StoresService.reloadStores();
+  });
+
+  $rootScope.$on('drag-start-item', (event, args) => {
+    dragBox = document.getElementById('item-drag-box');
+    vm.item = args.item;
+    dragBox.style.top = `${args.element.target.getBoundingClientRect().top - dragBox.offsetHeight}px`;
+    $rootScope.$digest();
+  });
+
+  $rootScope.$on('drag-stop-item', () => {
+    dragBox.style.top = '-200px';
+    vm.item = null;
+    $rootScope.$digest();
   });
 }