+++ src/app/destinyTrackerApi/reviewDataCache.js
@@ -133,8 +133,7 @@ class ReviewDataCache {
 
     if (matchingItem.reviews) {
       matchingItem.reviews = _.reject(matchingItem.reviews, { isReviewer: true });
-    }
-    else {
+    } else {
       matchingItem.reviews = [];
     }