+++ src/app/farming/d2farming.service.js
@@ -1,4 +1,5 @@
 import _ from 'underscore';
+import { REP_TOKENS } from './rep-tokens';
 
 /**
  * A service for "farming" items by moving them continuously off a character,
@@ -11,7 +12,8 @@ export function D2FarmingService($rootScope,
                         $interval,
                         toaster,
                         $i18next,
-                        D2BucketsService) {
+                        D2BucketsService,
+                        dimSettingsService) {
   'ngInject';
 
   let intervalId;
@@ -54,7 +56,7 @@ export function D2FarmingService($rootScope,
             if (vaultSpaceLeft <= 1) {
               // If we're down to one space, try putting it on other characters
               const otherStores = _.select(D2StoresService.getStores(),
-                                            (store) => !store.isVault && store.id !== store.id);
+                                            (s) => !s.isVault && s.id !== store.id);
               const otherStoresWithSpace = _.select(otherStores, (store) => store.spaceLeftForItem(item));
 
               if (otherStoresWithSpace.length) {
@@ -84,7 +86,7 @@ export function D2FarmingService($rootScope,
     makeRoomForItems: function(store) {
       return getMakeRoomBuckets().then((makeRoomBuckets) => {
         // If any category is full, we'll move one aside
-        const itemsToMove = [];
+        let itemsToMove = [];
         makeRoomBuckets.forEach((makeRoomBucket) => {
           const items = store.buckets[makeRoomBucket.id];
           if (items.length > 0 && items.length >= makeRoomBucket.capacity) {
@@ -108,6 +110,10 @@ export function D2FarmingService($rootScope,
             }
           }
         });
+
+        if (dimSettingsService.farming.moveTokens) {
+          itemsToMove = itemsToMove.concat(store.items.filter((i) => REP_TOKENS.has(i.hash)));
+        }
 
         if (itemsToMove.length === 0) {
           return $q.resolve();