+++ src/app/item-review/item-review.component.js
@@ -2,16 +2,22 @@ import _ from 'underscore';
 import template from './item-review.html';
 import './item-review.scss';
 
-function ItemReviewController(dimSettingsService, dimDestinyTrackerService, $scope) {
+function ItemReviewController(dimSettingsService, dimDestinyTrackerService, $scope, $rootScope) {
   'ngInject';
 
   const vm = this;
   vm.canReview = dimSettingsService.allowIdPostToDtr;
   vm.canCreateReview = (vm.canReview && vm.item.owner);
   vm.submitted = false;
-  vm.hasUserReview = vm.item.userRating;
-  vm.expandReview = vm.item.isLocallyCached;
+  vm.hasUserReview = ((vm.item.userRating) || (vm.item.userVote));
+  vm.expandReview = ((vm.item.isLocallyCached) && (vm.item.userVote !== 0));
   vm.toggledFlags = [];
+
+  vm.isCollapsed = false;
+
+  vm.toggleChart = function() {
+    vm.isCollapsed = !vm.isCollapsed;
+  };
 
   vm.procon = false; // TODO: turn this back on..
   vm.aggregate = {
@@ -25,21 +31,26 @@ function ItemReviewController(dimSettingsService, dimDestinyTrackerService, $sco
 
   vm.toggleEdit = function() {
     vm.expandReview = !vm.expandReview;
+
+    if ((vm.item.userVote === 1) ||
+        (vm.item.userVote === -1)) {
+      vm.item.userVote = 0;
+      vm.reviewBlur();
+    }
   };
 
   vm.clickReview = function(reviewId) {
-    const review = _.find(vm.item.writtenReviews, { reviewId: reviewId });
+    const review = this.findReview(reviewId);
 
     if (review.isReviewer) {
       vm.editReview();
-    }
-    else if (!review.isHighlighted) {
+    } else if (!review.isHighlighted) {
       vm.openFlagContext(reviewId);
     }
   };
 
   vm.openFlagContext = function(reviewId) {
-    const review = _.find(vm.item.writtenReviews, { reviewId: reviewId });
+    const review = this.findReview(reviewId);
 
     if ((review.isReviewer) || (review.isHighlighted)) {
       return;
@@ -58,14 +69,58 @@ function ItemReviewController(dimSettingsService, dimDestinyTrackerService, $sco
     vm.toggledFlags.splice(toggledReviewIndex);
   };
 
+  vm.findReview = function(reviewId) {
+    if (vm.item.destinyVersion === 1) {
+      return _.find(vm.item.writtenReviews, { reviewId: reviewId });
+    } else {
+      return _.find(vm.item.writtenReviews, { id: reviewId });
+    }
+  };
+
   vm.editReview = function(reviewId) {
-    const review = _.find(vm.item.writtenReviews, { reviewId: reviewId });
+    const review = this.findReview(reviewId);
 
     if (!review.isReviewer) {
       return;
     }
 
     vm.expandReview = true;
+
+    if (review.voted) {
+      vm.item.userVote = review.voted;
+    }
+  };
+
+  vm.totalReviews = 0;
+
+  vm.reviewLabels = [5, 4, 3, 2, 1];
+
+  vm.getReviewData = function() {
+    if (!vm.item.writtenReviews) {
+      return [];
+    }
+
+    const labels = vm.reviewLabels;
+
+    const mapData = _.map(labels, (label) => {
+      const matchingReviews = _.where(vm.item.writtenReviews, { rating: label });
+      const highlightedReviews = _.where(matchingReviews, { isHighlighted: true });
+
+      return matchingReviews.length + (highlightedReviews.length * 4);
+    });
+
+    vm.totalReviews = mapData.reduce((sum, cur) => { return sum + cur; }, 0);
+
+    return mapData;
+  };
+
+  vm.reviewData = vm.getReviewData();
+
+  vm.shouldDrawChart = function() {
+    vm.reviewData = vm.getReviewData();
+
+    return ((vm.reviewData.length > 0) &&
+            (_.some(vm.reviewData, (item) => { return item > 0; })));
   };
 
   vm.submitReview = function() {
@@ -90,12 +145,36 @@ function ItemReviewController(dimSettingsService, dimDestinyTrackerService, $sco
   };
 
   vm.reportReview = function(reviewId) {
-    const review = _.find(vm.item.writtenReviews, { reviewId: reviewId });
+    const review = this.findReview(reviewId);
 
     dimDestinyTrackerService.reportReview(review);
   };
 
   vm.toUserReview = function(item) {
+    if (vm.item.destinyVersion === 1) {
+      return this.toDestinyOneUserReview(item);
+    }
+
+    return this.toDestinyTwoUserReview(item);
+  };
+
+  vm.toDestinyTwoUserReview = function(item) {
+    const userVote = item.userVote;
+    const review = item.userReview;
+    const pros = item.userReviewPros;
+    const cons = item.userReviewCons;
+
+    const userReview = {
+      voted: userVote,
+      review: review,
+      pros: pros,
+      cons: cons
+    };
+
+    return userReview;
+  };
+
+  vm.toDestinyOneUserReview = function(item) {
     const newRating = item.userRating;
     const review = item.userReview;
     const pros = item.userReviewPros;
@@ -122,12 +201,28 @@ function ItemReviewController(dimSettingsService, dimDestinyTrackerService, $sco
     dimSettingsService.save();
   });
 
+  $rootScope.$on('dim-item-reviews-fetched', () => {
+    vm.reviewData = vm.getReviewData();
+  });
+
   vm.valueChanged = function() {
     vm.canReview = dimSettingsService.allowIdPostToDtr;
 
     if (vm.canReview) {
       dimDestinyTrackerService.getItemReviews(vm.item);
     }
+  };
+
+  vm.setUserVote = function(userVote) {
+    if (vm.item.userVote === userVote) {
+      vm.item.userVote = 0;
+    } else {
+      vm.item.userVote = userVote;
+    }
+
+    vm.expandReview = (vm.item.userVote !== 0);
+
+    vm.reviewBlur();
   };
 }